Simplify cli based tests
Avoid relying on brittle dynamic imports and console monkey patching
